requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.5
decorator==4.4.2
distlib==0.3.8
easyocr==1.7.1
elastic-transport==8.15.0
elasticsearch==8.14.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-gpu==1.7.2
fastapi==0.112.1
ffmpeg-python==0.2.0
filelock==3.15.4
Flask-Assets==2.1.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.1.1
future==1.0.0
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==3.1.0a0
grpcio==1.66.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.24.6
hypercorn==0.14.3
hyperframe==5.2.0
identify==2.6.0
idna==2.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.12.0
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
libretranslatepy==2.1.1
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.6.2
nodeenv==1.9.1
numba==0.60.0
numpy==1.23.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
open_clip_torch==2.26.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==1.5.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
preshed==3.0.9
priority==2.0.0
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-bidi==0.6.0
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-multipart==0.0.9
pytz==2024.1
pyvi==0.1.1
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==2.12.0
redis==3.5.3
regex==2022.9.13
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sklearn-crfsuite==0.5.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.2
sympy==1.13.2
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.8.10
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.64.1
traitlets==5.14.3
transformers==4.44.2
translate==3.6.1
triton==2.3.1
typer==0.12.5
typing_extensions==4.12.2
underthesea==1.3.5a3
underthesea_core==0.0.5a2
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webassets==2.0
websockets==13.0
Werkzeug==2.3.3
wrapt==1.14.1
wsproto==1.2.0
zipp==3.20.0
